# openapi.json
{
  "components": {
    "schemas": {
...
      "AgentCreate": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigCreate"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional starting config; recorded (dedup) and activated in the same call"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
      "AgentRead": {
        "properties": {
          "active_config_cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Config Cid"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_session_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Session Uuid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "uuid",
          "name",
          "created_at",
          "active_config_cid",
          "current_session_uuid"
        ],
        "title": "AgentRead",
        "type": "object"
      },
      "AgentRecordCreate": {
        "description": "One round of this agent's work; identity and session come from the path\nand the agent's state. ``actor_id`` defaults to the agent's name,\n``step_id`` to the ``span_id``.",
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "artifact_cids": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Cids",
            "type": "array"
          },
          "business_object_keys": {
            "items": {
              "type": "string"
            },
            "title": "Business Object Keys",
            "type": "array"
          },
          "config_cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "hex CID of the config used; defaults to the agent's active one",
            "title": "Config Cid"
          },
          "input_context": {
            "title": "Input Context",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "prev": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev"
          },
          "sequence": {
            "default": 0,
            "title": "Sequence",
            "type": "integer"
          },
          "span_id": {
            "description": "runtime span handle, assigned at span start",
            "title": "Span Id",
            "type": "string"
          },
          "step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "transformation": {
            "title": "Transformation",
            "type": "string"
          },
          "used_sessions": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Used Sessions",
            "type": "array"
          }
        },
        "required": [
          "span_id",
          "input_context",
          "agent_output",
          "transformation"
        ],
        "title": "AgentRecordCreate",
        "type": "object"
      },
      "AgentUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AgentUpdate",
        "type": "object"
      },
...
      "ConfigCreate": {
        "properties": {
          "llm_config": {
            "additionalProperties": true,
            "title": "Llm Config",
            "type": "object"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "tools": {
            "items": {},
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "system_prompt"
        ],
        "title": "ConfigCreate",
        "type": "object"
      },
...
      "DUTRead": {
        "properties": {
          "agent_cid": {
            "title": "Agent Cid",
            "type": "string"
          },
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "artifact_cids": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Cids",
            "type": "array"
          },
          "business_object_keys": {
            "items": {
              "type": "string"
            },
            "title": "Business Object Keys",
            "type": "array"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "config_cid": {
            "title": "Config Cid",
            "type": "string"
          },
          "input_context": {
            "title": "Input Context",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "span_id": {
            "title": "Span Id",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "span_id",
          "business_object_keys",
          "sequence",
          "input_context",
          "agent_output",
          "occurred_at",
          "agent_cid",
          "config_cid",
          "artifact_cids"
        ],
        "title": "DUTRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LineageRead": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "derived_from": {
            "items": {
              "type": "string"
            },
            "title": "Derived From",
            "type": "array"
          },
          "dut": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dut"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "prev": {
            "items": {
              "type": "string"
            },
            "title": "Prev",
            "type": "array"
          },
          "session_uuid": {
            "format": "uuid",
            "title": "Session Uuid",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "transformation": {
            "title": "Transformation",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "session_uuid",
          "actor_id",
          "step_id",
          "transformation",
          "occurred_at",
          "prev",
          "derived_from",
          "dut"
        ],
        "title": "LineageRead",
        "type": "object"
      },
      "NewSessionCreate": {
        "properties": {
          "session_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "stable session key; defaults to a new uuid4",
            "title": "Session Uuid"
          },
          "step_id": {
            "default": "session",
            "title": "Step Id",
            "type": "string"
          }
        },
        "title": "NewSessionCreate",
        "type": "object"
      },
...
      "RecordRead": {
        "properties": {
          "dut": {
            "$ref": "#/components/schemas/DUTRead"
          },
          "node": {
            "$ref": "#/components/schemas/LineageRead"
          }
        },
        "required": [
          "dut",
          "node"
        ],
        "title": "RecordRead",
        "type": "object"
      },
      "SessionCreate": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "session_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "stable session key; defaults to a new uuid4",
            "title": "Session Uuid"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "transformation": {
            "default": "session-start",
            "title": "Transformation",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "step_id"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
...
      "SessionSummary": {
        "properties": {
          "frontier": {
            "items": {
              "type": "string"
            },
            "title": "Frontier",
            "type": "array"
          },
          "last_occurred_at": {
            "format": "date-time",
            "title": "Last Occurred At",
            "type": "string"
          },
          "nodes": {
            "title": "Nodes",
            "type": "integer"
          },
          "session_uuid": {
            "format": "uuid",
            "title": "Session Uuid",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "session_uuid",
          "nodes",
          "started_at",
          "last_occurred_at",
          "frontier"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "aits",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agents/": {
      "get": {
        "operationId": "agents_list_all",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentRead"
                  },
                  "title": "Response Agents List All",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "agents_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "agents"
        ]
      }
    },
    "/agents/{agent_uuid}/": {
      "get": {
        "operationId": "agents_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "agents"
        ]
      },
      "patch": {
        "operationId": "agents_rename",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/agents/{agent_uuid}/record/": {
      "post": {
        "description": "Record one round of this agent's work into its current session,\nstarting a session if it has none. Concurrent operations on one agent\nshould use ``POST /lineage/{session_uuid}/record/`` with explicit sessions\ninstead of sharing this pointer.",
        "operationId": "agents_record",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record",
        "tags": [
          "agents"
        ]
      }
    },
    "/agents/{agent_uuid}/sessions/": {
      "post": {
        "description": "Start a fresh session for this agent and point its current-session\nref at it; subsequent ``record`` calls chain there. Returns the session's\ngenesis node.",
        "operationId": "agents_new_session",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Session",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/lineage/": {
      "get": {
        "description": "Enumerate sessions, most recently active first. Includes sub-sessions\n(e.g. subagent work); their parents reach them via ``derived_from``.",
        "operationId": "lineage_list_sessions",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Lineage List Sessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "lineage"
        ]
      },
      "post": {
        "operationId": "lineage_start_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "lineage"
        ]
      }
    },
...
  }
}